requirements.txt
contourpy==1.2.1
cycler==0.12.1
filelock==3.14.0
fonttools==4.51.0
fsspec==2024.3.1
importlib_resources==6.4.0
# install==1.3.5
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.4
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pillow==10.3.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
scikit-learn==1.4.2
scipy==1.13.0
six==1.16.0
sklearn==0.0
sympy==1.12
threadpoolctl==3.5.0
torch==2.2.2
torchvision==0.17.2
typing_extensions==4.11.0
zipp==3.18.1
